Apply safer percentage in pallet-accounts calculations

+++ pallets/accounts/src/dummy_data.rs
@@ -27,14 +27,14 @@ pub fn get_ticket_asset_id() -> u64 {
 }
 
 #[cfg(test)]
-pub fn get_initial_percentage_tickets() -> u128 {
+pub fn get_initial_percentage_tickets() -> u8 {
 	use crate::mock::Test;
 	use frame_support::traits::TypedGet;
 	<Test as Config>::InitialPercentageTickets::get()
 }
 
 #[cfg(test)]
-pub fn get_initial_percentage_nova() -> u128 {
+pub fn get_initial_percentage_nova() -> u8 {
 	use crate::mock::Test;
 	use frame_support::traits::TypedGet;
 	<Test as Config>::InitialPercentageNova::get()
